Add the sr25519 implement and test

+++ tcx-substrate/src/lib.rs
@@ -11,7 +11,7 @@ pub(crate) const ACCOUNT_INDEX_FLAG: u8 = 0xff;
 pub(crate) const SIGNATURE_TYPE_ED25519: u8 = 0x00;
 pub(crate) const SIGNATURE_TYPE_SR25519: u8 = 0x01;
 pub(crate) const SIGNED_EXTRINSIC_V4: u8 = 0x84;
-pub(crate) const PAYLOAD_HASH_THRESHOLD: int = 256;
+pub(crate) const PAYLOAD_HASH_THRESHOLD: usize = 256;
 
 #[cfg(test)]
 mod tests {
